requirements.txt
Flask
python-dotenv
pyjwt
bcrypt
psycopg2-binary
gunicorn
